update refresh
Nu word de user opgehaald uit de app (als het goed is)
