+++ src/App.jsx
@@ -1,42 +1,68 @@
+import { useState, useEffect } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import moment from 'moment';
-import 'moment/dist/locale/nl';
-import './App.css'
+import { doc, getDoc } from "firebase/firestore";
+import { db } from "./firebase";
 
+import Header from './components/Header';
 import Login from './pages/Login';
-import ProtectedRoute from './pages/ProtectedRoute';
-import Voorpagina from './pages/Voorpagina'
+import Voorpagina from './pages/Voorpagina';
+import Profiel from './pages/Profiel';
 import Ziekmelden from './pages/Ziekmelden';
-import GebruikerToevoegen from './pages/GebruikerToevoegen';
 import VerlofAanvraag from './pages/VerlofAanvraag';
-import Profiel from './pages/Profiel'
-import AuditOverzicht from './pages/AuditOverzicht'
+import GebruikerToevoegen from './pages/GebruikerToevoegen';
+import AuditOverzicht from './pages/AuditOverzicht';
+import ProtectedRoute from './pages/ProtectedRoute';
 
 function App() {
+  const [user, setUser] = useState(null);
+
+  const fetchUser = async () => {
+    const gebruikerId = localStorage.getItem("userId");
+    if (!gebruikerId) return;
+
+    try {
+      const gebruikerDoc = await getDoc(doc(db, "user", gebruikerId));
+      if (!gebruikerDoc.exists()) return;
+
+      const gebruikerData = gebruikerDoc.data();
+
+      let rolNaam = "Onbekende rol";
+      if (gebruikerData.rol_id) {
+        const rolDoc = await getDoc(gebruikerData.rol_id);
+        if (rolDoc.exists()) rolNaam = rolDoc.data().rolNaam || "Onbekende rol";
+      }
+
+      setUser({ ...gebruikerData, rol: rolNaam });
+
+    } catch (error) {
+      console.error("Fout bij ophalen gebruiker:", error);
+    }
+  };
+
+  useEffect(() => {
+    fetchUser();
+  }, []);
+
+  // Redirect logica
   const isLoggedIn = localStorage.getItem("isLoggedIn");
   const rol = localStorage.getItem("rol");
-
-  if (!isLoggedIn && window.location.pathname !== "/") {
-    window.location.href = "/";
-  }
-
+  if (!isLoggedIn && window.location.pathname !== "/") window.location.href = "/";
   if (isLoggedIn && !rol) {
     localStorage.removeItem("isLoggedIn");
     window.location.href = "/";
   }
 
-  moment.locale('nl');
-
   return (
     <BrowserRouter>
+      <Header user={user} refreshUser={fetchUser} />
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/ziekmelden" element={<Ziekmelden />} />
         <Route path="/verlofaanvraag" element={<VerlofAanvraag />} />
         <Route path="/audit-overzicht" element={<AuditOverzicht />} />
         <Route path="/gebruiker-registratie" element={<GebruikerToevoegen />} />
+        <Route path="/profiel/:userId" element={<Profiel refreshUser={fetchUser} />} />
 
-        {/* Voorpagina's beschermd per rol */}
         <Route path="/officemanager/voorpagina" element={
           <ProtectedRoute allowedRoles={["officemanager"]}>
             <Voorpagina />
@@ -54,8 +80,6 @@ function App() {
             <Voorpagina />
           </ProtectedRoute>
         }/>
-
-        <Route path="/profiel/:userId" element={<Profiel />} />
 
         <Route path="*" element={<Navigate to="/" replace />} />
       </Routes>
