+++ src/components/Header.jsx
@@ -1,47 +1,10 @@
 import GeoprofsLogoWit from '../media/GeoprofsLogoWit.png';
 import { useNavigate } from 'react-router-dom';
-import { useState, useEffect } from 'react';
-import { db } from "../firebase";
-import { doc, getDoc } from "firebase/firestore";
+import { useState } from 'react';
 
-function Header() {
+function Header({ user, refreshUser }) {
   const navigate = useNavigate();
   const [profielMenu, setProfielMenu] = useState(false);
-  const [gebruiker, setGebruiker] = useState(null);
-
-  useEffect(() => {
-    const haalGebruikerOp = async () => {
-      const gebruikerId = localStorage.getItem("userId");
-      if (!gebruikerId) return;
-
-      try {
-        //Haalt de gebruiker op
-        const gebruikerDoc = await getDoc(doc(db, "user", gebruikerId));
-        if (!gebruikerDoc.exists()) return;
-
-        const gebruikerData = gebruikerDoc.data();
-
-        let rolNaam = "Onbekende rol";
-
-        if (gebruikerData.rol_id) {
-          const rolDoc = await getDoc(gebruikerData.rol_id);
-          if (rolDoc.exists()) {
-            rolNaam = rolDoc.data().rolNaam || "Onbekende rol";
-          }
-        }
-
-        setGebruiker({
-          ...gebruikerData,
-          rol: rolNaam
-        });
-
-      } catch (error) {
-        console.error("Fout bij ophalen gebruiker:", error);
-      }
-    };
-
-    haalGebruikerOp();
-  }, []);
 
   const handleUitloggen = () => {
     localStorage.removeItem("userId");
@@ -55,7 +18,10 @@ function Header() {
       <header className="h-[10%] w-full bg-[#E8641C] flex items-center">
         <div className="w-[1%]" />
         <div className="w-[50%] h-[80%]">
-          <a href={`/${gebruiker?.rol?.toLowerCase().replaceAll(" ", "")}/voorpagina`} className="inline-block w-[40%] h-[90%]">
+          <a
+            href={`/${user?.rol?.toLowerCase().replaceAll(" ", "")}/voorpagina`}
+            className="inline-block w-[40%] h-[90%]"
+          >
             <img src={GeoprofsLogoWit} alt="Geoprofs Logo" className="h-full w-auto" />
           </a>
         </div>
@@ -64,20 +30,13 @@ function Header() {
           <button className="flex cursor-pointer" onClick={() => setProfielMenu(!profielMenu)}>
             <div className="h-full w-auto mr-5">
               <div className="flex w-full h-[55%] text-xl items-end justify-end text-white">
-                <p>
-                  {/* hier heb ik aan gewerkt in de profiel-database feature, was niet zo slim */}
-                  {gebruiker 
-                    ? `${gebruiker.voornaam} ${gebruiker.achternaam}` 
-                    : "Bezig met ophalen..."
-                  }
-                </p>
+                <p>{user ? `${user.voornaam} ${user.achternaam}` : "Bezig met ophalen..."}</p>
               </div>
               <div className="flex w-full h-[45%] text-sm justify-end text-white">
-                {/* hier heb ik aan gewerkt in de profiel-database feature, was niet zo slim */}
-                <p>{gebruiker?.rol || "Onbekende rol"}</p>
+                <p>{user?.rol || "Onbekende rol"}</p>
               </div>
             </div>
-            <img src="" alt="Profiel Foto" className='h-full aspect-square bg-[#fff] rounded-full'/>
+            <img src="" alt="Profiel Foto" className="h-full aspect-square bg-[#fff] rounded-full" />
           </button>
         </div>
 
@@ -88,22 +47,22 @@ function Header() {
         <div className="h-[60%] absolute top-0 right-4 pointer-events-none">
           <div className="h-[20%]" />
           <div className="w-[200px] bg-white border-2 border-[#D0D0D0] p-[5px] pointer-events-auto rounded-lg shadow-md">
-            
-            <button 
+
+            <button
               className="h-[40px] w-full cursor-pointer border-2 border-[#D0D0D0] rounded-[15px] mb-[5px]"
               onClick={() => navigate(`/profiel/${localStorage.getItem("userId")}`)} >
               Profiel
             </button>
 
-            {gebruiker?.rol === "Office Manager" && (
-              <button 
+            {user?.rol === "Office Manager" && (
+              <button
                 className="h-[40px] w-full cursor-pointer border-2 border-[#D0D0D0] rounded-[15px] mb-[5px]"
                 onClick={() => navigate(`/audit-overzicht`)} >
                 Audit Overzicht
               </button>
             )}
 
-            <button 
+            <button
               className="h-[40px] w-full cursor-pointer border-2 rounded-[15px] bg-[#DF121B] text-white"
               onClick={handleUitloggen} >
               Uitloggen
